requirements.txt
asgiref==3.10.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.7.4
channels==4.3.1
channels-redis==4.3.0
charset-normalizer==3.3.2
cryptography==44.0.0
celery==5.4.0
Django==5.0.3
django-allauth==0.63.6
django-celery-beat==2.7.0
django-celery-results==2.6.0
django-anymail==13.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-crontab==0.7.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
Faker==12.0.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mixer==7.2.2
msgpack==1.1.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-html==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pika==1.3.2
redis==7.0.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
#rpds-py==0.19.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
# beautifulsoup4==4.12.3
# bs4==0.0.2
reportlab
flask
django-extensions